Merges origin/main (which now includes the retired-model-removal and
wire-format-endpoint-fix PRs) into this branch. ROADMAP.md conflicted
because both this branch and the retired-model-removal PR appended a
milestone entry after #62; resolved by keeping both narratively in
order (dated cleanup entry, then '63. Wire Format Conformance Fix
Pass', then '64. TC18 RCP GA', renumbered from '63').

Two integration-only defects surfaced by combining this branch's
avtp::ByteBusId widening (uint8_t -> uint16_t, an 11-bit wire field)
with code neither this branch nor the retired-model-removal branch
touched on its own:

- tests/test_rcp.cpp (added by the retired-model-removal PR, which
  predates this branch's ByteBusId widening) asserted
  ByteBusId == uint8_t via static_assert; updated to assert the type
  is wide enough for the real 11-bit range instead.
- include/rcp/discovery.hpp's encode_discovery_request took
  sequence_num as uint16_t and assigned it into NtscfHeader::sequence_num,
  which this branch narrowed to uint8_t (sequence_num is an 8-bit wire
  field, not 16) as part of the same header fix. Neither this file nor
  its test was in this branch's original diff, so the mismatch only
  appeared once merged; narrowed the parameter type to match.

Full local build (C++17, Ninja) and ctest (52/52 tests) pass on the
merged tree.

+++ include/rcp/discovery.hpp
@@ -111,7 +111,7 @@ inline acf::AcfMessageInfo make_discovery_request(uint8_t transaction_num,
 // produces a TSCF-headed frame, so there is no way to misuse it into
 // building the kind of request decode_discovery_request below must drop.
 inline std::vector<uint8_t> encode_discovery_request(const avtp::StreamId& stream_id,
-                                                       uint16_t sequence_num,
+                                                       uint8_t sequence_num,
                                                        uint8_t transaction_num,
                                                        uint16_t read_size = kDiscoveryDefaultReadSize) {
     const auto info = make_discovery_request(transaction_num, read_size);
